Refactor key management to use ulong for key IDs/results
Refactored all key ID, key result, and start position types from int to ulong across the key management system. Updated method signatures, bitwise operations, and related logic in KeyService, MatrixKeyGenerator, KeyServiceOptions, and tests. This change improves scalability, supports larger key spaces, and ensures consistency in handling key-related identifiers.

+++ Mrbr.Service.KeyManager/Services/IKeyService.cs
@@ -11,93 +11,93 @@ public interface IKeyService {
     /// </summary>
     /// <param name="keyId">Output key identifier</param>
     /// <returns>Key material (Block keys) or empty (Matrix keys)</returns>
-    ReadOnlyMemory<char> GenerateKey(out int keyId);
+    ReadOnlyMemory<char> GenerateKey(out ulong keyId);
     /// <summary>
     /// Generate a fixed-length 16-byte key from the Key Manager (selects from 0-255 keys).
     /// </summary>
-    byte[] GenerateKey128(out int keyId, KeyDerivationOptions options = default);
+    byte[] GenerateKey128(out ulong keyId, KeyDerivationOptions options = default);
     /// <summary>
     /// Generate a fixed-length 24-byte key from the Key Manager (selects from 0-255 keys).
     /// </summary>
-    byte[] GenerateKey192(out int keyId, KeyDerivationOptions options = default);
+    byte[] GenerateKey192(out ulong keyId, KeyDerivationOptions options = default);
     /// <summary>
     /// Generate a fixed-length 32-byte key from the Key Manager (selects from 0-255 keys).
     /// </summary>
-    byte[] GenerateKey256(out int keyId, KeyDerivationOptions options = default);
+    byte[] GenerateKey256(out ulong keyId, KeyDerivationOptions options = default);
     /// <summary>
     /// Generate a fixed-length key from the Key Manager into a caller-provided destination span (selects from 0-255 keys).
     /// Destination length must be 16, 24, or 32 bytes.
     /// </summary>
-    void GenerateKeyBytes(Span<byte> destination, out int keyId, KeyDerivationOptions options = default);
+    void GenerateKeyBytes(Span<byte> destination, out ulong keyId, KeyDerivationOptions options = default);
     /// <summary>
     /// Generate a fixed-length key from the Key Manager (selects from 0-255 keys).
     /// keySizeInBytes must be 16, 24, or 32.
     /// </summary>
-    byte[] GenerateKeyBytes(int keySizeInBytes, out int keyId, KeyDerivationOptions options = default);
+    byte[] GenerateKeyBytes(int keySizeInBytes, out ulong keyId, KeyDerivationOptions options = default);
     /// <summary>
     /// Generate a fixed-length 16-byte key from the Key Manager into a caller-provided destination span (selects from 0-255 keys).
     /// </summary>
-    void GenerateKey128(Span<byte> destination, out int keyId, KeyDerivationOptions options = default);
+    void GenerateKey128(Span<byte> destination, out ulong keyId, KeyDerivationOptions options = default);
     /// <summary>
     /// Generate a fixed-length 24-byte key from the Key Manager into a caller-provided destination span (selects from 0-255 keys).
     /// </summary>
-    void GenerateKey192(Span<byte> destination, out int keyId, KeyDerivationOptions options = default);
+    void GenerateKey192(Span<byte> destination, out ulong keyId, KeyDerivationOptions options = default);
     /// <summary>
     /// Generate a fixed-length 32-byte key from the Key Manager into a caller-provided destination span (selects from 0-255 keys).
     /// </summary>
-    void GenerateKey256(Span<byte> destination, out int keyId, KeyDerivationOptions options = default);
+    void GenerateKey256(Span<byte> destination, out ulong keyId, KeyDerivationOptions options = default);
     /// <summary>
     /// Get the Key text from the Source Keys using the Key Id (0-255).
     /// </summary>
     /// <param name="keyId">Key identifier (0-255)</param>
     /// <returns>Key material</returns>
-    ReadOnlyMemory<char> GetKey(int keyId);
+    ReadOnlyMemory<char> GetKey(ulong keyId);
     /// <summary>
     /// Get a fixed-length 16-byte key from the Key Manager using a specific key ID (0-255).
     /// </summary>
-    byte[] GetKey128(int keyId, KeyDerivationOptions options = default);
+    byte[] GetKey128(ulong keyId, KeyDerivationOptions options = default);
     /// <summary>
     /// Get a fixed-length 24-byte key from the Key Manager using a specific key ID (0-255).
     /// </summary>
-    byte[] GetKey192(int keyId, KeyDerivationOptions options = default);
+    byte[] GetKey192(ulong keyId, KeyDerivationOptions options = default);
     /// <summary>
     /// Get a fixed-length 32-byte key from the Key Manager using a specific key ID (0-255).
     /// </summary>
-    byte[] GetKey256(int keyId, KeyDerivationOptions options = default);
+    byte[] GetKey256(ulong keyId, KeyDerivationOptions options = default);
     /// <summary>
     /// Get a fixed-length key from the Key Manager into a caller-provided destination span using a specific key ID (0-255).
     /// Destination length must be 16, 24, or 32 bytes.
     /// </summary>
-    void GetKeyBytes(int keyId, Span<byte> destination, KeyDerivationOptions options = default);
+    void GetKeyBytes(ulong keyId, Span<byte> destination, KeyDerivationOptions options = default);
     /// <summary>
     /// Get a fixed-length key from the Key Manager using a specific key ID (0-255).
     /// keySizeInBytes must be 16, 24, or 32.
     /// </summary>
-    byte[] GetKeyBytes(int keyId, int keySizeInBytes, KeyDerivationOptions options = default);
+    byte[] GetKeyBytes(ulong keyId, ulong keySizeInBytes, KeyDerivationOptions options = default);
     /// <summary>
     /// Get a fixed-length 16-byte key from the Key Manager into a caller-provided destination span using a specific key ID (0-255).
     /// </summary>
-    void GetKey128(int keyId, Span<byte> destination, KeyDerivationOptions options = default);
+    void GetKey128(ulong keyId, Span<byte> destination, KeyDerivationOptions options = default);
     /// <summary>
     /// Get a fixed-length 24-byte key from the Key Manager into a caller-provided destination span using a specific key ID (0-255).
     /// </summary>
-    void GetKey192(int keyId, Span<byte> destination, KeyDerivationOptions options = default);
+    void GetKey192(ulong keyId, Span<byte> destination, KeyDerivationOptions options = default);
     /// <summary>
     /// Get a fixed-length 32-byte key from the Key Manager into a caller-provided destination span using a specific key ID (0-255).
     /// </summary>
-    void GetKey256(int keyId, Span<byte> destination, KeyDerivationOptions options = default);
+    void GetKey256(ulong keyId, Span<byte> destination, KeyDerivationOptions options = default);
     /// <summary>
     /// Check if the Key Id exists in the Key Manager (0-255).
     /// </summary>
     /// <param name="keyId">Key identifier (0-255)</param>
     /// <returns>True if key exists</returns>
-    bool ContainsKey(int keyId);
+    bool ContainsKey(ulong keyId);
     /// <summary>
     /// Delete the Key Id from the Key Manager (0-255).
     /// </summary>
     /// <param name="keyId">Key identifier (0-255)</param>
     /// <returns>True if key was deleted</returns>
-    bool DeleteKey(int keyId);
+    bool DeleteKey(ulong keyId);
     /// <summary>
     /// Delete all Keys from the Key Manager (all 256 slots).
     /// </summary>
@@ -107,8 +107,8 @@ public interface IKeyService {
     /// Get a random Key Id from the available keys (0-255).
     /// </summary>
     /// <returns>Random key ID</returns>
-    int GetRandomKeyId();
+    ulong GetRandomKeyId();
 
-    Task<(ReadOnlyMemory<char> Key, int Id)> GenerateKeyAsync();
-    Task<ReadOnlyMemory<char>> GetKeyAsync(int keyId);
+    Task<(ReadOnlyMemory<char> Key, ulong Id)> GenerateKeyAsync();
+    Task<ReadOnlyMemory<char>> GetKeyAsync(ulong keyId);
 }
